+++ src/main/java/gtclassic/GTOreDict.java
@@ -80,6 +80,7 @@ public class GTOreDict {
 		// My Stuff
 		OreDictionary.registerOre("craftingToolMortar", new ItemStack(GTItems.flintMortar, 1, OreDictionary.WILDCARD_VALUE));
 		OreDictionary.registerOre("craftingToolMortar", new ItemStack(GTItems.ironMortar, 1, OreDictionary.WILDCARD_VALUE));
+		OreDictionary.registerOre("craftingToolDuctTape", new ItemStack(GTItems.ductTape, 1, OreDictionary.WILDCARD_VALUE));
 		OreDictionary.registerOre("circuitMaster", (GTMaterialGen.get(GTItems.circuitEnergy, 1)));
 		OreDictionary.registerOre("circuitElite", (GTMaterialGen.get(GTItems.circuitData, 1)));
 		OreDictionary.registerOre("circuitData", (GTMaterialGen.get(GTItems.chipData, 1)));